Implementação dos arquivos swagger.json dos novos microsserviços e adição de validação no método delete das atividades

+++ API Monolítica - AP1/servico_atividades_notas/controllers/atividades_controller.py
@@ -148,6 +148,9 @@ def delete_atividade(atividade_id):
     if not atividade:
         return jsonify({'error': 'Atividade não encontrada!'}), 404
 
+    if atividade.notas:
+        return jsonify({'error': 'Não é possível deletar essa atividade porque tem notas vinculadas!'}), 400
+
     db.session.delete(atividade)
     db.session.commit()
     return jsonify({'message': 'Atividade removida com sucesso!'}), 200
